Split CodeGenerator helpers into named classes
